requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
bcrypt==5.0.0
    # via paramiko
certifi==2026.2.25
    # via requests
cffi==2.0.0
    # via
    #   cryptography
    #   pynacl
charset-normalizer==3.4.5
    # via requests
contourpy==1.3.3
    # via matplotlib
cryptography==46.0.5
    # via
    #   google-auth
    #   paramiko
cycler==0.12.1
    # via matplotlib
dataclasses-json==0.5.14
    # via doccano-client
doccano-client==1.2.8
    # via -r requirements.in
fonttools==4.62.1
    # via matplotlib
google-api-core==2.30.0
    # via google-api-python-client
google-api-python-client==2.193.0
    # via -r requirements.in
google-auth==2.49.1
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
google-auth-httplib2==0.3.0
    # via
    #   -r requirements.in
    #   google-api-python-client
google-auth-oauthlib==1.3.0
    # via -r requirements.in
googleapis-common-protos==1.73.0
    # via google-api-core
httplib2==0.31.2
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.11
    # via requests
invoke==2.2.1
    # via paramiko
joblib==1.5.3
    # via scikit-learn
kiwisolver==1.5.0
    # via matplotlib
marshmallow==3.26.2
    # via dataclasses-json
matplotlib==3.10.8
    # via seaborn
mypy-extensions==1.1.0
    # via typing-inspect
numpy==2.4.3
    # via
    #   contourpy
    #   matplotlib
    #   pandas
    #   scikit-learn
    #   scipy
    #   seaborn
oauthlib==3.3.1
    # via requests-oauthlib
packaging==26.0
    # via
    #   marshmallow
    #   matplotlib
pandas==3.0.1
    # via
    #   -r requirements.in
    #   seaborn
paramiko==4.0.0
    # via -r requirements.in
pillow==12.1.1
    # via matplotlib
proto-plus==1.27.1
    # via google-api-core
protobuf==6.33.6
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
pyasn1==0.6.3
    # via pyasn1-modules
pyasn1-modules==0.4.2
    # via google-auth
pycparser==3.0
    # via cffi
pydantic==1.10.26
    # via doccano-client
pynacl==1.6.2
    # via paramiko
pyparsing==3.3.2
    # via
    #   httplib2
    #   matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-dotenv==1.1.1
    # via -r requirements.in
requests==2.32.5
    # via
    #   doccano-client
    #   google-api-core
    #   requests-oauthlib
    #   requests-toolbelt
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
requests-toolbelt==0.9.1
    # via doccano-client
scikit-learn==1.8.0
    # via -r requirements.in
scipy==1.17.1
    # via scikit-learn
seaborn==0.13.2
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
threadpoolctl==3.6.0
    # via scikit-learn
typing-extensions==4.15.0
    # via
    #   pydantic
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
uritemplate==4.2.0
    # via google-api-python-client
urllib3==1.26.20
    # via
    #   -r requirements.in
    #   requests
